+++ KeyboardKit/Extensions/UITextDocumentProxy+Delete.swift
@@ -5,6 +5,7 @@
 //  Created by Daniel Saidi on 2019-07-02.
 //  Copyright © 2019 Daniel Saidi. All rights reserved.
 //
+//  TODO: Unit test this extension
 
 import UIKit