+++ KeyboardKitExampleKeyboard/KeyboardViewController.swift
@@ -37,11 +37,13 @@ import KeyboardKit
  move these parts to `KeyboardInputViewController` and a new
  api for working with autocomplete.
  
- TODO: `textDidChange` does not trigger when the user types.
- Until I have found a way to detect text input, I have added
- a temporary timer that triggers each second and requests an
- autocomplete operation for the current word. I have created
- an issue to solve this in another way.
+ **IMPORTANT** `textWillChange` and `textDidChange` does not
+ trigger when a user types and text is sent to the proxy. It
+ however works when the text cursor changes its position, so
+ I therefore use a (hopefully temporary) hack, by starting a
+ timer that triggers each second and moves the cursor. Since
+ this is a nasty hack, it may have yet to be discovered side
+ effects. If so, please let me know.
  
  */
 class KeyboardViewController: KeyboardInputViewController {
@@ -75,12 +77,12 @@ class KeyboardViewController: KeyboardInputViewController {
     
     override func selectionWillChange(_ textInput: UITextInput?) {
         super.selectionWillChange(textInput)
-        resetAutocompleteSuggestions()
+        autocompleteToolbar.reset()
     }
     
     override func selectionDidChange(_ textInput: UITextInput?) {
         super.selectionDidChange(textInput)
-        requestAutocompleteSuggestions()
+        autocompleteToolbar.reset()
     }
     
     
@@ -123,7 +125,7 @@ class KeyboardViewController: KeyboardInputViewController {
     }
     
     private func resetAutocompleteSuggestions() {
-        autocompleteToolbar.update(with: [])
+        autocompleteToolbar.reset()
     }
 }